Rename object test to serde util test
